requirements.txt
cryptography
hashids
pymysql
pyjwt
sqlalchemy
fastapi